Added VG_(get_fnname_if_entry) which can be used by skins to instrument the entry to a function, if it has not had its symbol stripped.
Beefed up lackey to use it to count number of calls to _dl_runtime_resolve().
Also added BB and x86_instr counters to lackey, and calculate some ratios at
the end.

Also for --trace-codegen now detecting function entry BBs and printing out the
name in the BB header if found.