Back out rev 1.9 (optimised %eflags save/restore).  It isn't quite right.
Here's a code sequence illustrating the problem.  The conditional jump
at the end evidently goes the wrong way sometimes, and the program
goes off into outer space soon after.

        0x4017F6F7:  addl %ebx,%ecx

          12: GETL      %ECX, t10
          13: ADDL      %EBX, t10  (-wOSZACP)
          14: PUTL      t10, %ECX
          15: INCEIPo   $2

        0x4017F6F9:  decl %eax

          16: GETL      %EAX, t12
          17: DECL      t12  (-wOSZAP)
          18: PUTL      t12, %EAX
          19: INCEIPo   $1

        0x4017F6FA:  jnb-8 0x4017F710

          20: Jnbo      $0x4017F710  (-rOSZACP)
          21: JMPo      $0x4017F6FC

Look carefully at the annotation on # 17.  Then look in the Intel docs
and see what flag(s) the Jnb (not-below) condition consults.  Bwaha!
It consults the carry flag.

The generated code for 17 (renamed to 12 after some NOP removal, I
guess) is

          12: DECL      %edx  (-wOSZAP)         [---d--]
                 42: 4A
                decl    %edx
                 43: 9C 8F 45 20
                pushfl ; popl 32(%ebp)

viz, we do the decl, and then copy the real machine's %eflags into
%EFLAGS.  Unfortunately this copies the real carry flag into the
simulated one, rather than leaving the simulated one alone.

So the principle is that it's only safe to omit the initial
%EFLAGS->%eflags move prior to the insn if the insn writes _all_ the
flags, and in this case it doesn't.

(after further consideration ...)

At first it seems tempting to play games with subset checks, ie if an
insn writes a _subset_ of the flags, we'd better copy sim'd to real
flags before the insn.

Problem with that is that the D (direction) flag, which specifies the
direction that rep-prefix string ops travel, is part of the "normal"
flag set.  So it is conceivable, although highly unlikely, that an app
could

- set the D flag
- do something like   ADD (-wOSZACP)
- use the D flag

then we'd still have to do an %EFLAGS->%eflags copy prior to the ADD,
in order to ensure the simulated D flag is preserved.

So it seems to me that it's pretty much impossible to preserve
absolute correctness and do any better than the vg_from_ucode.c rev
1.8 scheme.

Backing out rev 1.9.  We can just as easily restore it from cvs if a
better solution is arrived at.