Moved the MALLOCLIKE and FREELIKE client requests out of memcheck.h, and into valgrind.h.  Although these requests are not implemented by the core, they can be implemented by skins that track heap blocks, eg. Memcheck, Annelid, Massif. This is in preparation for committing Massif to the repository.
I think I managed to make the change in a binary-compatible way.  The only
inconvenience for users is that if they have a client program compiled with the
old requests in, Valgrind will abort with an explanatory message that tells
them to recompile.  Once they've done that (no changes to their program are
required), it works again.

I even updated the docs.