To get 32-bit programs working on Opteron, VG_(valgrind_end) was recently changed to name the last byte in Valgrind's section, rather than one past the last byte.  This was because the last byte is 0xffffffff, and so one past gave 0x0, which screwed things up.
However, when this change was made, all the places where VG_(valgrind_end) is
used weren't adjusted appropriately.  So this commit makes those adjustments.
It also renames the variable as VG_(valgrind_last), which makes the difference
between it and the other VG_(*_end) variables much clearer.

MERGED FROM HEAD