Change VG_(nuke_all_threads) to disassociate the the stacks of the threads being killed from the threads rather than marking them as inaccessible.
This should fix the problem with the environment (and other data from the
stacks of other threads) causing warnings after a fork. I believe that
VG_(nuke_all_threads) is only called in places where this is the behaviour
that we want or where it doesn't matter because we're about to exit anyway.

I believe this fixes bug 85625 but I'm still waiting for confirmation.