Make async-style syscalls work on ppc64, by using rt_sigprocmask instead of sigprocmask.
In the process, discover that error handling for
ML_(do_syscall_for_client_WRK) on all platforms has always been
broken, in the sense that the sigprocmasks (which are important) could
silently fail.  This commit fixes that up too (only on ppc64-linux at
the moment, so all other platforms are probably broken now).