Enhancements to debuginfo segment-lookup functions:
Change VG_(seginfo_sect_kind) so it also produces an object file
name as well as a VgSectKind.  Adjust users accordingly.  Make
use of this in Memcheck's address-describer.

Remove VG_(seginfo_sect_kind_name) and replace it with
VG_(pp_SectKind), so as to independently have a way to print
VgSectKinds without doing the address lookup.