The value of AT_BASE should be the offset between where the ELF interpreter expected to be loaded (as expressed in the ELF headers) and where it was actually loaded, and not (as valgrind was doing) the absolute value of the load address for the interpreter.
Note that when prelink is not in use the two are normally the same, as the
intpreter (like all shared libraries) is normally linked with a zero load
address. When prelinked that is no longer true.

With that fixed, the hack to patch out AT_BASE to avoid confusing gdb on
systems where prelink is in use is no longer needed.

Fixes BZ#329612